# contract/openapi.yaml
openapi: 3.1.0
info:
  title: AgentMux Public Integration API
  description: >
    Public integration contract for AgentMux. Stability tiers (stable / beta /
    internal) are defined in CONTRACT.md; only stable and beta endpoints are
    described here. The OpenAI-compatible surface (/v1/responses, /v1/files)
    follows the upstream OpenAI Responses specification and is intentionally
    not duplicated in this document.
  version: "1.3"
  x-contract-version: "1.3"
servers:
  - url: http://127.0.0.1:8765
    description: Local AgentMux daemon (default listen address)
security:
  - bridgeToken: []
...
paths:
...
  /api/v1/tenancy/register:
    post:
      tags: [tenancy]
      operationId: registerTenant
      summary: Create a tenant and its first token in one call
      description: >
        Self-service creation with no approval step and no existing credential.
        Tenant names are unique: registering an existing name returns 409
        rather than rotating that tenant's credentials, so this path can only
        mint fresh identities. A new tenant starts with no accessible private
        resources; the administrator can grant or assign Agents, channels,
        triggers and Providers later.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Tenant name to register. Must be unused.
                kind:
                  type: string
                  enum: [app, web, service]
      responses:
        "200":
          description: The tenant and its first token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantRegistration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: A tenant with this name already exists
  /api/v1/tenancy/self:
    get:
      tags: [tenancy]
      operationId: getTenancySelf
      summary: Read back the calling credential's identity and scope
      responses:
        "200":
          description: Caller identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantSelf"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bridgeToken:
      type: http
      scheme: bearer
      description: >
        The [bridge] token from config.toml. Required on /api/* and /v1/*
        whenever bridge.enabled = true. A valid agentmux_console session
        cookie is an equivalent credential for browser-embedded Console use.
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid bridge token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Tenant:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string }
        name: { type: string }
        kind:
          type: string
          enum: [app, web, service]
        status:
          type: string
          enum: [active, disabled]
        note: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TenantRegistration:
      type: object
      required: [tenant, token]
      properties:
        tenant: { $ref: '#/components/schemas/Tenant' }
        token:
          type: string
          description: >
            The tenant's long-lived bearer token. Returned exactly once; it is
            stored only as a hash and cannot be recovered later.
        prefix:
          type: string
          description: Displayed truncation of the token, for later reference.
    TenantSelf:
      type: object
      required: [admin]
      properties:
        admin:
          type: boolean
          description: True when the caller holds administrator credentials.
        tenant_id: { type: string }
        tenant: { type: string }
        kind: { type: string }
        status: { type: string }
...